added Spring Security roles :sparkles:
