+++ src/main/resources/db/migration/V1__Create_Eshop_tables.sql
@@ -1,3 +1,12 @@
+-- ROLES
+--
+CREATE TABLE IF NOT EXISTS roles
+(
+    role_id SERIAL      NOT NULL,
+    name    varchar(20) NOT NULL,
+    PRIMARY KEY (role_id)
+);
+
 -- USERS
 --
 CREATE TABLE IF NOT EXISTS users
@@ -9,7 +18,9 @@ CREATE TABLE IF NOT EXISTS users
     birth_date date NOT NULL,
     balance    NUMERIC(10, 2) DEFAULT '0.00',
     role_id INTEGER,
-    PRIMARY KEY (user_id)
+    PRIMARY KEY (user_id),
+    CONSTRAINT fk_role_id FOREIGN KEY (role_id) REFERENCES roles (role_id)
+
 );
 
 -- CATEGORIES
@@ -58,13 +69,4 @@ CREATE TABLE IF NOT EXISTS carts
     user_id    INTEGER,
     PRIMARY KEY (cart_id),
     CONSTRAINT fk_user_id FOREIGN KEY (user_id) REFERENCES users (user_id)
-);
-
--- ROLES
---
-CREATE TABLE IF NOT EXISTS roles
-(
-    role_id SERIAL      NOT NULL,
-    name    varchar(20) NOT NULL,
-    PRIMARY KEY (role_id)
 );
