+++ src/main/resources/templates/include/topnav.html
@@ -9,10 +9,9 @@
         <a class="fa badge fa-5x" href="/logout">&#xf08b Log out</a>
         <a class="fa badge fa-5x" href="/profile">&#xf007 Profile</a>
         <a class="fa badge-cart fa-5x" href="/cart" th:value="${session.cartDto.quantity}">&#xf07a Cart</a>
-        <a class="fa badge fa-5x" href="/admin">Admin</a>
     </div>
     <div sec:authorize="hasRole('ADMIN')">
-        <a class="fa badge fa-5x" href="/admin">Admin Role</a>
+        <a class="fa badge fa-5x" href="/admin">Admin</a>
     </div>
     <div class="search-container">
         <form class="search-form" th:action="@{/product/search}">
